configurator: Adjust camera-lights synchronization
Delay between sending a frame and grabbing a picture has been
increased from 30ms to 100ms to account for various delays, including
the time it takes for the LED light to turn on/off or camera delay.

This number was chosen after manually testing various values
and checking how many "off" pictures were not actually off,
but contained the previous light still on. 100ms turned out to
be the smallest delay where all the "off" pictures were consistently off.