visualizer: Fix camera perspective
This commit makes two changes to how the scene and the camera are set up.

First, all Z coordinates of the points are reversed, since we use
a left-handed coordinate system (X increases to the right, Y increases
to the top and Z increases away from you), bevy uses a right-handed
system (Z increases towards you). This means the tree was mirrored
in the visualizer.

Second, the position of the camera is changed from (3, 0, 0) to (0, 0, 3),
so that the camera is in front of the tree, not to the right of it.